NO-JIRA: Refactor EnsureNoCrashingPods function
Refactored EnsureNoCrashingPods as a first step to use an standard interface for all Ensure functions
