+++ test/e2e/nodepool_machineconfig_test.go
@@ -79,6 +79,10 @@ func (mc *NodePoolMachineconfigRolloutTest) BuildNodePoolManifest(defaultNodepoo
 }
 
 func (mc *NodePoolMachineconfigRolloutTest) Run(t *testing.T, nodePool hyperv1.NodePool, nodes []corev1.Node) {
+	var (
+		EnsureFuncNoCrashingPods e2eutil.EnsureFunc = e2eutil.EnsureNoCrashingPods
+	)
+
 	g := NewWithT(t)
 
 	// MachineConfig Actions
@@ -144,7 +148,10 @@ func (mc *NodePoolMachineconfigRolloutTest) Run(t *testing.T, nodePool hyperv1.N
 	eventuallyDaemonSetRollsOut(t, ctx, mc.hostedClusterClient, len(nodes), np, ds)
 	g.Expect(nodePool.Status.Replicas).To(BeEquivalentTo(len(nodes)))
 
-	e2eutil.EnsureNoCrashingPods(t, ctx, mc.mgmtClient, mc.hostedCluster)
+	EnsureFuncNoCrashingPods(t, ctx, &e2eutil.TestParams{
+		MgmtClient:    mc.mgmtClient,
+		HostedCluster: mc.hostedCluster,
+	})
 	e2eutil.EnsureAllContainersHavePullPolicyIfNotPresent(t, ctx, mc.mgmtClient, mc.hostedCluster)
 	e2eutil.EnsureHCPContainersHaveResourceRequests(t, ctx, mc.mgmtClient, mc.hostedCluster)
 	e2eutil.EnsureNoPodsWithTooHighPriority(t, ctx, mc.mgmtClient, mc.hostedCluster)
